Google Music Lyrics now apart of Baraka Lyrics
This addition implements Google Music Lyrics to Baraka Lyrics.

**Caution** this is limited to USA only with-in the code I added a
check to see if the geo of the external ip in use is in the US thanks
to geoplugin.

The limitation is due to Google only having Google Music Lyrics in the
US only.

**Note with Google Music Lyrics not every song is available with
lyrics, Google only adds lyrics to select songs so if you get “No
lyrics Available” or a nil that is the reason.**

This is a work in-progress but for now enjoy!